# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nova Hub",
...
    "contact": {
      "name": "Nova Hub",
      "url": "https://github.com/yourusername/nova-hub"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/leagues/{league_id}/packets/{filename}": {
      "put": {
        "tags": [
          "Packets"
        ],
        "summary": "Upload Packet",
        "description": "Upload a game packet to the hub using PUT with raw body\n\n**Path Parameters:**\n- `league_id`: League identifier with game type (e.g., \"555B\" for BRE, \"555F\" for FE)\n- `filename`: Packet filename (e.g., \"555B0201.001\")\n\n**Request Body:** Raw file data (application/octet-stream)\n\n**Packet Filename Format:** `<league><game><source><dest>.<seq>`\n- League: 3 digits (e.g., 555)\n- Game: B (BRE) or F (FE)\n- Source: 2-digit hex BBS index (e.g., 02)\n- Dest: 2-digit hex BBS index (e.g., 01)\n- Seq: 3-digit sequence (000-999)\n\n**Example:** `555B0201.001` = BRE League 555, from BBS 02 to BBS 01, seq 1\n\n**Authentication:** Requires Bearer token\n\n**Example:**\n```bash\ncurl -X PUT \"https://hub.example.com/api/v1/leagues/555B/packets/555B0201.001\" \\\n  -H \"Authorization: Bearer YOUR_TOKEN\" \\\n  -H \"Content-Type: application/octet-stream\" \\\n  --data-binary \"@555B0201.001\"\n```",
        "operationId": "upload_packet_api_v1_leagues__league_id__packets__filename__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "league_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{3}[BF]$",
              "title": "League Id"
            }
          },
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Packets"
        ],
        "summary": "Download Packet",
        "description": "Download a specific packet file\n\n**Path Parameters:**\n- `league_id`: League identifier with game type (e.g., \"555B\" for BRE, \"555F\" for FE)\n- `filename`: Packet filename (e.g., \"555B0102.001\")\n\n**Returns:** Binary packet file\n\n**Description:**\nDownloads a packet that is destined for your BBS. The packet is automatically\nmarked as downloaded after successful retrieval.\n\n**Authentication:** Requires Bearer token and authorization for destination BBS\n\n**Example:**\n```bash\ncurl -X GET \"https://hub.example.com/api/v1/leagues/555B/packets/555B0102.001\" \\\n  -H \"Authorization: Bearer YOUR_TOKEN\" \\\n  -o 555B0102.001\n```",
        "operationId": "download_packet_api_v1_leagues__league_id__packets__filename__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "league_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{3}[BF]$",
              "title": "League Id"
            }
          },
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leagues/{league_id}/packets": {
      "get": {
        "tags": [
          "Packets"
        ],
        "summary": "List Available Packets",
        "description": "List packets available for download by this client\n\n**Path Parameters:**\n- `league_id`: League identifier with game type (e.g., \"555B\" for BRE, \"555F\" for FE)\n\n**Query Parameters:**\n- `unread`: Set to `true` to only show packets not yet downloaded (default: false)\n\n**Returns:** List of packets destined for this client's BBS\n\n**Authentication:** Requires Bearer token\n\n**Example:**\n```bash\n# Get all packets\ncurl -X GET \"https://hub.example.com/api/v1/leagues/555B/packets\" \\\n  -H \"Authorization: Bearer YOUR_TOKEN\"\n\n# Get only unread packets\ncurl -X GET \"https://hub.example.com/api/v1/leagues/555B/packets?unread=true\" \\\n  -H \"Authorization: Bearer YOUR_TOKEN\"\n```",
        "operationId": "list_packets_api_v1_leagues__league_id__packets_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "league_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{3}[BF]$",
              "title": "League Id"
            }
          },
          {
            "name": "unread",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Filter to only unread (not downloaded) packets",
              "default": false,
              "title": "Unread"
            },
            "description": "Filter to only unread (not downloaded) packets"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leagues/{league_id}/nodelist": {
      "get": {
        "tags": [
          "Packets"
        ],
        "summary": "Download Nodelist",
        "description": "Download the latest nodelist file for a league\n\n**Path Parameters:**\n- `league_id`: League identifier with game type (e.g., \"555B\" for BRE, \"555F\" for FE)\n\n**Returns:** Binary nodelist file (BRNODES.xxx or FENODES.xxx)\n\n**Description:**\nThe nodelist contains information about all BBS nodes in the league.\nThese files are generated by the hub and should be placed in your game directory.\n\n**Authentication:** Requires Bearer token and league membership\n\n**Example:**\n```bash\ncurl -X GET \"https://hub.example.com/api/v1/leagues/555B/nodelist\" \\\n  -H \"Authorization: Bearer YOUR_TOKEN\" \\\n  -o BRNODES.555\n```",
        "operationId": "download_nodelist_api_v1_leagues__league_id__nodelist_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "league_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{3}[BF]$",
              "title": "League Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/token"
          }
        }
      }
    }
  }
}